refactor(section): 인수 테스트 리팩터링
